speech/pyproject.toml
[project]
name = "speech"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "elevenlabs>=2.9.2",
    "python-dotenv>=1.1.1",
]
